Phase3_Smart_Contract/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bitarray==2.9.3
certifi==2025.4.26
charset-normalizer==3.4.2
cytoolz==0.12.3
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.7.0
hexbytes==0.3.1
idna==3.10
ipfshttpclient==0.8.0a2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lru-dict==1.3.0
multiaddr==0.0.9
multidict==6.4.4
netaddr==1.3.0
parsimonious==0.8.1
propcache==0.3.2
protobuf==3.19.5
pycryptodome==3.23.0
python-dotenv==1.0.0
referencing==0.36.2
requests==2.31.0
rlp==2.0.1
rpds-py==0.25.1
six==1.17.0
toolz==1.0.0
typing_extensions==4.14.0
urllib3==2.4.0
varint==1.0.2
web3==5.31.4
websockets==9.1
yarl==1.20.1
